requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
pyautogui>=0.9.53
opencv-python>=4.5.0
pillow>=9.0.0
pyperclip>=1.8.0
pyyaml>=6.0
